BusModule can't open itself if we're in OSGi.
We allow the extender to open the BusModule so that we can be sure it's
visible to the other tracker inside of HandlerRegistry and when the
subscription messages come in they'll be seen. Before there was basically
a race between sending the subscribe message (during child startup) and
having a Handler be visible to the registry. It would always lose because
the handler wasn't visible until after BusModule was registered.